requirements.txt
aiohttp==2.2.5
aiopg==0.13.1
async-timeout==1.3.0
chardet==3.0.4
marshmallow==2.13.6
multidict==3.1.3
peewee==2.10.1
peewee-async==0.5.7
psycopg2==2.7.3.1
python-rapidjson==0.2.4
webargs==1.8.1
yarl==0.12.0
